#33 Changed return type of SchedulerParent::getTotalRequests
- Completed MPI_Scheduler implementation as this one already had the member variable

+++ GemPBA/MPI_Modules/MPI_Scheduler_Centralized.hpp
@@ -122,7 +122,7 @@ namespace gempba {
             spdlog::info("\n \n \n");
         }
 
-        double getTotalRequests() const override {
+        size_t getTotalRequests() const override {
             return totalRequests;
         }
